+++ R/apply-pgs.R
@@ -143,7 +143,11 @@ validate.phenotype.data.input <- function(phenotype.data, phenotype.analysis.col
 #'
 #' \strong{Missing Genotype Handling}
 #'
-#' Missing genotypes are handled by three methods:
+#' VCF genotype data are matched to PGS data by chromosome, position, and effect allele. If a SNP cannot be matched by genomic coordinate,
+#' an attempt is made to match by rsID (if available). If a SNP from the PGS weight data is not found in the VCF data after these two matching attempts,
+#' it is considered a cohort-wide missing variant.
+#'
+#' Missing genotypes (in individual samples) among successfully matched variants are handled by three methods:
 #'
 #' \code{none}: Missing genotype dosages are excluded from the PGS calculation.
 #' This is equivalent to assuming that all missing genotypes are homozygous for the non-effect allele, resulting in a dosage of 0.
@@ -158,13 +162,13 @@ validate.phenotype.data.input <- function(phenotype.data, phenotype.analysis.col
 #' where \emph{k} is a PGS component variant that is missing in between 1 and n-1 individuals in the cohort and \emph{P} = ploidy = 2
 #' This dosage calculation holds under assumptions of Hardy-Weinberg equilibrium.
 #' By default, the effect allele frequency is calculated from the provided VCF data.
-#' For variants that are missing in all individuals, dosage is assumed to be zero (homozygous non-reference) for all individuals.
+#' For variants that are missing in all individuals (cohort-wide), dosage is assumed to be zero (homozygous non-reference) for all individuals.
 #' An external allele frequency can be provided in the \code{pgs.weight.data} as a column named \code{allelefrequency_effect} and by setting \code{use.external.effect.allele.frequency} to \code{TRUE}.
 #'
 #' \strong{Multiallelic Site Handling}
 #'
-#' VCF genotype data are matched to PGS data by chromosome, position, and effect allele. If a PGS weight file provides weights for multiple effect alleles, the appropriate dosage is calculated for the
-#' alleles that each individual carries. It is assumed that multiallelic variants are encoded in the same row in the VCF data. This is known as "merged" format. Split multiallelic sites are not accepted.
+#' If a PGS weight file provides weights for multiple effect alleles, the appropriate dosage is calculated for the alleles that each individual carries.
+#' It is assumed that multiallelic variants are encoded in the same row in the VCF data. This is known as "merged" format. Split multiallelic sites are not accepted.
 #' VCF data can be formatted to merged format using external tools for VCF file manipulation.
 #'
 #' @examples
@@ -293,18 +297,22 @@ apply.polygenic.score <- function(
         );
 
     ### Start Missing Genotype Handling ###
-
+    variant.id <- paste(merged.vcf.with.pgs.data$CHROM, merged.vcf.with.pgs.data$POS, merged.vcf.with.pgs.data$effect_allele, sep = ':');
     if ('mean.dosage' %in% missing.genotype.method) {
         # calculate dosage to replace missing genotypes
         if (use.external.effect.allele.frequency) {
-            missing.genotype.dosage <- convert.allele.frequency.to.dosage(allele.frequency = pgs.weight.data$allelefrequency_effect);
-            names(missing.genotype.dosage) <- paste(pgs.weight.data$CHROM, pgs.weight.data$POS, pgs.weight.data$effect_allele, sep = ':');
+            missing.genotype.dosage <- data.frame(dosage = convert.allele.frequency.to.dosage(allele.frequency = merged.vcf.with.pgs.data$allelefrequency_effect));
+            missing.genotype.dosage$variant.id <- variant.id;
+            # remove duplicated rows
+            missing.genotype.dosage <- missing.genotype.dosage[!duplicated(missing.genotype.dosage$variant.id), ];
+            # convert to named vector
+            missing.genotype.dosage <- setNames(missing.genotype.dosage$dosage, missing.genotype.dosage$variant.id);
             # identify missing genotypes
             # this method includes variants that are missing in all Indivs, these are all replaced with the same mean dosage
             missing.genotype.row.index <- which(is.na(merged.vcf.with.pgs.data$dosage));
             } else {
             # create sample by variant dosage matrix
-            variant.id <- paste(merged.vcf.with.pgs.data$CHROM, merged.vcf.with.pgs.data$POS, merged.vcf.with.pgs.data$effect_allele, sep = ':');
+
             dosage.matrix <- get.variant.by.sample.matrix(
                 long.data = merged.vcf.with.pgs.data,
                 variant.id = variant.id,